backend/requirements.txt
Flask==2.3.3
Flask-Cors==4.0.0
pymongo==4.6.0
requests==2.31.0
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.2
xgboost==2.0.2
tensorflow==2.13.0
keras==2.13.1
textblob==0.17.1
PyJWT==2.8.0
gunicorn==21.2.0
python-dotenv==1.0.0
joblib==1.3.2